Format source code PolicyEngineClient
Formatting and sonar issues.

Looks like this is all code used for demonstrating how to call
the API's. Uses all main so actually no JUnit is being called.

Unfortunately there are no associated test files for it. Would be
interesting to see if adding a JUnit to call it with some test
code would help to improve overall coverage. Future work.

Issue-ID: POLICY-2145
Change-Id: I0d0f6d8898076a92d342553f45f70a56fd9d49f6
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>