Update patch set 2
Patch Set 2:

(2 comments)

we've configured lombok so that sonar doesn't count instructions generated by lombok.  That should reduce the denominator thereby improving the coverage count.  That being said, I'm not sure if the effect is the same with ONAP sonar since it counts lines instead of instructions.  (And, of course, eliminating 100 lines isn't even noticeable in the statistics :-( )

Patch-set: 2
Label: Code-Review=0