Update patch set 2
Patch Set 2: -Code-Review

Perhaps Praveen can take a look at this and determine what's up.  It appears from the output that the ubuntu container in which the build is run doesn't recognize its own name when it's used in a URL by junit tests.  Not sure why this suddenly started happening, but wondering if the container is running out of memory (it has 8GB, so that would be odd)

Patch-set: 2