Update patch set 1
Patch Set 1: Code-Review+2

> Patch Set 1: Code-Review+1
> 
> I guess at some point we need to put back the snapshot value back in at least some of the files that were touched when preparing to release in https://gerrit.onap.org/r/#/c/policy/engine/+/104431/.   Probably we should do it before we forget.

I disagree - only if we need to release those artifacts again do we have to go back and point to the RELEASED artifacts. This doesn't need to go back to SNAPSHOT, we aren't in the middle of development. And frankly, I'm hoping this is the last time we ever have to release this repo ever.

Patch-set: 1
Reviewer: Gerrit User 57 <57@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+2