Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Is this all that was needed?  I noticed that ./ONAP-SDK-APP/pom.xml also has a reference to tomcat, but it's an older version.

I think this may be enough for the lab problems.   I saw the ONAP-SDK, but afraid to touch it, seems it is fairly dated, and don't want to bring another variable to the installation.   I don't know what the ONAP-SDK is but I don't think it plays a role on getting tomcat installed.

Patch-set: 1